// case_study/my-app/src/types.ts
// custom datatype definitions here
export type ProductType = {
    productId: number;
    productName: string;
    productImage: string;
    productPrice: number;
    productSalePrice: number;
    productStock: number;
  };

  export type UserType = {
    kind:         string;
    localId:      string;
    email:        string;
    displayName:  string;
    idToken:      string;
    registered:   boolean;
    refreshToken: string;
    expiresIn:    string;
}

export type Cart= {
  product: ProductType;
  count: number;
}